Feature/ISO-645

Approved-by: Will Wolff-Myren <wwolff-myren@learning.com>
Approved-by: David Rick <drick@learning.com>
Approved-by: David Hogue <dhogue@learning.com>
Approved-by: Dave Sanderling <dsanderling@learning.com>
Approved-by: Robert Haber <rhaber@learning.com>

+++ src/Learning.MessageQueue/Messages/RetryableRedisSubscription.cs
@@ -105,7 +105,7 @@ namespace Learning.MessageQueue.Messages
             var retryData = await _messageQueueRepository.GetRetryData(@event).ConfigureAwait(false);
 
             // exponential backoff
-            var mpow = (int)Math.Pow(2, retryData.RetryCount);
+            var mpow = Math.Pow(2, retryData.RetryCount);
             var interval = Math.Min(mpow * RetryIntervalMinutes, RetryIntervalMaxMinutes);
             var intervalPassed = DateTimeOffset.UtcNow > retryData.LastRetryTime?.ToUniversalTime().AddMinutes(interval);
